Add the Isle of Bute Discovery Centre Cinema
This cinema is a small independent (one screen, 97 seats) and the
listings are on a Blogspot blog that is updated each month with a
hand-edited HTML table.

The "lxml" package has been added to the archiver to support the
correct parsing of XML Atom/RSS feeds in BeautifulSoup, such as
those used by Blogspot.