CLI: Add "--dummy-run" option to not make any changes to DB
The `--dummy-run` option runs the archiver as normal, but does not
save any of the results into the DB. This is useful in a number of
circumstances:

1. Development: during development it is essential to know that
   the code compiles and runs without errors, but before this is
   known, having partial or incomplete results saved in the DB
   is undesirable.

2. Testing: automated checks to ensure that there have been no
   breaking API changes are desirable, but such tests should not
   change a DB.

Remaining issues: the URL log that maintains a list of URLs that
have been accessed to perform the archiving is not affected by the
status of the `--dummy-run` flag. This is undesirable, however I
am not sure what the most elegent solution to fixing this would be.