Consts: allow the webdriver order to be configured
Allow the selection and preference of the webdriver (currently used
just for Vista Systems - Odeon, Curzon - but likely more in the
future) to be user configurable.

Two webdrivers, Chrome and Firefox, are implemented currently, and
the default configuration in consts.py is to attempt to use Firefox
first, and then try Chrome if Firefox is not available.

If none of the given webdrivers are available, a
CinemaArchiverException is thrown.