Relax rules for what counts as an invalid showing
Previously a row in the table was considered invalid only if both
the title and time were blank. This is unsuitable as sometimes
the title field is used to display a message about why no showing
is occuring.

Changing the rules to result in an invalid showing if either the
title or time fields are blank results in the correct behaviour.