Allow for advance month publications
Sometimes two months' worth of listings will be published at once,
as is the case of November and December 2022. The problem with this
was that the cinema archiver was configured to only consider the
first XML feed item, and so would miss the other month.

This has been re-configured to accept all the showings in the
feed that are in the future.