Add the Isle of Bute Discovery Centre Cinema (#34)
* Add the Isle of Bute Discovery Centre Cinema

This cinema is a small independent (one screen, 97 seats) and the
listings are on a Blogspot blog that is updated each month with a
hand-edited HTML table.

The "lxml" package has been added to the archiver to support the
correct parsing of XML Atom/RSS feeds in BeautifulSoup, such as
those used by Blogspot.

* Fix formatting issues

* Fix type error

* Relax rules for what counts as an invalid showing

Previously a row in the table was considered invalid only if both
the title and time were blank. This is unsuitable as sometimes
the title field is used to display a message about why no showing
is occuring.

Changing the rules to result in an invalid showing if either the
title or time fields are blank results in the correct behaviour.

* Allow for advance month publications

Sometimes two months' worth of listings will be published at once,
as is the case of November and December 2022. The problem with this
was that the cinema archiver was configured to only consider the
first XML feed item, and so would miss the other month.

This has been re-configured to accept all the showings in the
feed that are in the future.

Co-authored-by: David Ferguson <davidferguson@users.noreply.github.com>